% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive_fetch_measurements}
\alias{archive_fetch_measurements}
\title{fetch measurements from archive from a single box, and a single sensor}
\usage{
archive_fetch_measurements(box, sensor, fromDate, toDate, progress)
}
\description{
fetch measurements from archive from a single box, and a single sensor
}
