% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{build_archive_url}
\alias{build_archive_url}
\title{returns URL to fetch measurements from a sensor for a specific date,
based on `osem_archive_endpoint()`}
\usage{
build_archive_url(date, box, sensor)
}
\description{
returns URL to fetch measurements from a sensor for a specific date,
based on `osem_archive_endpoint()`
}
