% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_utils.R
\name{filter.osem_measurements}
\alias{filter.osem_measurements}
\title{Return rows with matching conditions, while maintaining class & attributes}
\usage{
\method{filter}{osem_measurements}(.data, ..., .dots)
}
\arguments{
\item{.data}{A osem_measurements data.frame to filter}

\item{...}{other arguments}

\item{.dots}{see corresponding function in package \code{\link{dplyr}}}
}
\description{
Return rows with matching conditions, while maintaining class & attributes
}
\seealso{
\code{\link[dplyr]{filter}}
}
