% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_generics.R
\name{filter.sensebox}
\alias{filter.sensebox}
\title{Return rows with matching conditions, while maintaining class & attributes}
\usage{
\method{filter}{sensebox}(.data, ..., .dots)
}
\arguments{
\item{.data}{A sensebox data.frame to filter}

\item{...}{other arguments}

\item{.dots}{see corresponding function in package \code{\link{dplyr}}}
}
\description{
Return rows with matching conditions, while maintaining class & attributes
}
\seealso{
\code{\link[dplyr]{filter}}
}
