% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_utils.R
\name{mutate.osem_measurements}
\alias{mutate.osem_measurements}
\title{Add new variables to the data, while maintaining class & attributes}
\usage{
\method{mutate}{osem_measurements}(.data, ..., .dots)
}
\arguments{
\item{.data}{A osem_measurements data.frame to mutate}

\item{...}{other arguments}

\item{.dots}{see corresponding function in package \code{\link{dplyr}}}
}
\description{
Add new variables to the data, while maintaining class & attributes
}
\seealso{
\code{\link[dplyr]{mutate}}
}
