% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_generics.R
\name{mutate.sensebox}
\alias{mutate.sensebox}
\title{Add new variables to the data, while maintaining class & attributes}
\usage{
mutate.sensebox(.data, ..., .dots)
}
\arguments{
\item{.data}{A sensebox data.frame to mutate}

\item{...}{other arguments}

\item{.dots}{see corresponding function in package \code{\link{dplyr}}}
}
\description{
Add new variables to the data, while maintaining class & attributes
}
\seealso{
\code{\link[dplyr]{mutate}}
}
