% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opensensmapr.R
\docType{package}
\name{opensensmapr}
\alias{opensensmapr}
\alias{opensensmapr-package}
\alias{opensensmapr-package}
\title{opensensmapr: Work with sensor data from opensensemap.org}
\description{
The opensensmapr package provides three categories functions:
\enumerate{
  \item retrieval of senseBoxes
  \item retrieval of measurements
  \item general stats about the openSenseMap database
}
}
\section{Retrieving senseBox metadata}{

TODO
}

\section{Retrieving measurements}{

TODO
}

\section{Retrieving statistics}{

TODO
}

\section{Working with spatial data from openSenseMap}{

TODO
}

\seealso{
Useful links:
\itemize{
  \item \url{http://github.com/noerw/opensensmapR}
  \item Report bugs at \url{http://github.com/noerw/opensensmapR/issues}
}

}
