% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{osem_archive_endpoint}
\alias{osem_archive_endpoint}
\title{Returns the default endpoint for the archive *download*
While the front end domain is archive.opensensemap.org, file downloads
are provided via sciebo.}
\usage{
osem_archive_endpoint()
}
\description{
Returns the default endpoint for the archive *download*
While the front end domain is archive.opensensemap.org, file downloads
are provided via sciebo.
}
