% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measurement_utils.R
\name{osem_as_measurements}
\alias{osem_as_measurements}
\title{Converts a foreign object to an osem_measurements data.frame.}
\usage{
osem_as_measurements(x)
}
\arguments{
\item{x}{A data.frame to attach the class to}
}
\description{
Converts a foreign object to an osem_measurements data.frame.
}
