% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_utils.R
\name{osem_as_sensebox}
\alias{osem_as_sensebox}
\title{Converts a foreign object to a sensebox data.frame.}
\usage{
osem_as_sensebox(x)
}
\arguments{
\item{x}{A data.frame to attach the class to}
}
\description{
Converts a foreign object to a sensebox data.frame.
}
