% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00utils.R
\name{osem_as_sf}
\alias{osem_as_sf}
\title{Convert a \code{sensebox} or \code{osem_measurements} dataframe to an
\code{\link[sf]{st_sf}} object.}
\usage{
osem_as_sf(x, ...)
}
\arguments{
\item{x}{The object to convert}

\item{...}{maybe more objects to convert}
}
\value{
The object with an st_geometry column attached.
}
\description{
Convert a \code{sensebox} or \code{osem_measurements} dataframe to an
\code{\link[sf]{st_sf}} object.
}
