% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{osem_box_to_archivename}
\alias{osem_box_to_archivename}
\title{replace chars in box name according to archive script:
https://github.com/sensebox/osem-archiver/blob/612e14b/helpers.sh#L66}
\usage{
osem_box_to_archivename(box)
}
\arguments{
\item{box}{A sensebox data.frame}
}
\value{
character with archive identifier for each box
}
\description{
replace chars in box name according to archive script:
https://github.com/sensebox/osem-archiver/blob/612e14b/helpers.sh#L66
}
