% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts.R
\name{osem_counts}
\alias{osem_counts}
\title{Get count statistics of the openSenseMap Instance}
\usage{
osem_counts(endpoint = "https://api.opensensemap.org")
}
\arguments{
\item{endpoint}{The URL of the openSenseMap API}
}
\value{
A named \code{list} containing the counts
}
\description{
Provides information on number of senseBoxes, measurements, and measurements per minute.
}
\details{
Note that the API caches these values for 5 minutes.
}
\seealso{
\href{https://docs.opensensemap.org/#api-Misc-getStatistics}{openSenseMap API documentation (web)}
}
