% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{osem_endpoint}
\alias{osem_endpoint}
\title{Get the default openSenseMap API endpoint}
\usage{
osem_endpoint()
}
\value{
A character string with the HTTP URL of the openSenseMap API
}
\description{
Get the default openSenseMap API endpoint
}
