% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{osem_ensure_archive_available}
\alias{osem_ensure_archive_available}
\title{Check if the given openSenseMap archive endpoint is available}
\usage{
osem_ensure_archive_available(endpoint = osem_archive_endpoint())
}
\arguments{
\item{endpoint}{The archive base URL to check, defaulting to \code{\link{osem_archive_endpoint}}}
}
\value{
\code{TRUE} if the archive is available, otherwise \code{stop()} is called.
}
\description{
Check if the given openSenseMap archive endpoint is available
}
