% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{osem_measurements_archive}
\alias{osem_measurements_archive}
\alias{osem_measurements_archive.sensebox}
\title{Get day-wise measurements for a single box from the openSenseMap archive.}
\usage{
osem_measurements_archive(x, ...)

\method{osem_measurements_archive}{sensebox}(x, fromDate,
  toDate = fromDate, sensorFilter = ~T, progress = T)
}
\description{
This function is significantly faster than `osem_measurements()` for large
time-frames, as dayly CSV dumps for each sensor from
<archive.opensensemap.org> are used.
Note that the latest data available is from the previous day.
By default, data for all sensors of a box is fetched, but you can select a
subset with a `dplyr`-style NSE filter expression.
}
\section{Methods (by class)}{
\itemize{
\item \code{sensebox}: Get daywise measurements for one or
more sensors of a single box
}}

\examples{

\donttest{
  # fetch measurements for a single day
  box = osem_box('593bcd656ccf3b0011791f5a')
  m = osem_measurements_archive(box, as.POSIXlt('2018-09-13'))
  
  # fetch measurements for a date range and selected sensors
  sensors = ~ phenomenon \%in\% c('Temperatur', 'Beleuchtungsstärke')
  m = osem_measurements_archive(box, as.POSIXlt('2018-09-01'), as.POSIXlt('2018-09-30'), sensorFilter = sensors)
}
}
