% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{osem_archive_endpoint}
\alias{osem_archive_endpoint}
\title{Default endpoint for the archive download
front end domain is archive.opensensemap.org, but file download
is provided via sciebo}
\usage{
osem_archive_endpoint()
}
\description{
Default endpoint for the archive download
front end domain is archive.opensensemap.org, but file download
is provided via sciebo
}
