% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box.R
\name{osem_box}
\alias{osem_box}
\title{Get a single senseBox by its ID}
\usage{
osem_box(boxId, endpoint = "https://api.opensensemap.org")
}
\arguments{
\item{boxId}{A string containing a senseBox ID}

\item{endpoint}{The URL of the openSenseMap API instance}
}
\value{
A \code{sensebox data.frame} containing a box in each row
}
\description{
Get a single senseBox by its ID
}
\examples{
# get a specific box by ID
b = osem_box('593bcd656ccf3b0011791f5a')

}
\seealso{
\href{https://docs.opensensemap.org/#api-Measurements-findAllBoxes}{openSenseMap API documentation (web)}

\code{\link{osem_phenomena}}

\code{\link{osem_boxes}}
}
