% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/external_generics.R
\name{st_as_sf.osem_measurements}
\alias{st_as_sf.osem_measurements}
\title{Convert a \code{osem_measurements} dataframe to an \code{\link[sf]{st_sf}} object.}
\usage{
\method{st_as_sf}{osem_measurements}(x, ...)
}
\arguments{
\item{x}{The object to convert}

\item{...}{maybe more objects to convert}
}
\value{
The object with an st_geometry column attached.
}
\description{
Convert a \code{osem_measurements} dataframe to an \code{\link[sf]{st_sf}} object.
}
